+++ wrapper/plugwise/src/de/uniol/inf/is/odysseus/wrapper/plugwise/CRC16.java
@@ -55,7 +55,7 @@ public class CRC16 implements Checksum {
     @Override
 	public void update(byte[] b, int off, int len) {
         for (int i = off; i < off+len; i++)
-            update((int)b[i]);
+            update(b[i]);
     }
 
     @Override
